requirements.txt
# FastAPI 核心
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# 数据库
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
alembic==1.13.1

# Pydantic
pydantic==2.5.3
pydantic-settings==2.1.0

# Celery
celery==5.3.6
redis==5.0.1

# 认证
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0

# Azure OpenAI
openai==1.10.0

# 工具库
aiofiles==23.2.1
python-dateutil==2.8.2
numpy==1.26.3
pandas==2.1.4

# WebSocket
websockets==12.0

# 日志
structlog==24.1.0

# 测试
pytest==7.4.4
pytest-asyncio==0.23.3
httpx==0.26.0

# 开发工具
black==24.1.1
flake8==7.0.0
mypy==1.8.0
